feat: ensure DbDeleteResponse calls getQuestionRepository correctly

+++ src/data/usecases/delete-response/db-delete-response.ts
@@ -1,4 +1,5 @@
 import { DeleteResponse, DeleteResponseModel } from '../../../domain/usecases/delete-response'
+import { InvalidParamError } from '../../../presentation/errors/invalid-param-error'
 import { DeleteResponseRepository } from '../../protocols/delete-response-repository'
 import { GetQuestionRepository } from '../../protocols/get-question-repository'
 
@@ -15,6 +16,10 @@ export class DbDeleteResponse implements DeleteResponse {
   }
 
   async delete (response: DeleteResponseModel): Promise<Boolean> {
+    const questionExist = await this.getQuestionRepository.get(response.question_id)
+    if (!questionExist) {
+      throw new InvalidParamError('question_id')
+    }
     return await this.deleteResponseRepository.deleteResponse(response)
   }
 }
